create BulkEditSessionViewMixin
