+++ corehq/apps/data_cleaning/views/main.py
@@ -11,6 +11,7 @@ from django.utils.translation import gettext_lazy, gettext as _
 from corehq import toggles
 from corehq.apps.data_cleaning.models import BulkEditSession
 from corehq.apps.data_cleaning.tasks import commit_data_cleaning
+from corehq.apps.data_cleaning.views.mixins import BulkEditSessionViewMixin
 from corehq.apps.domain.decorators import login_and_domain_required
 from corehq.apps.hqwebapp.decorators import use_bootstrap5
 from corehq.apps.settings.views import BaseProjectDataView
@@ -39,7 +40,7 @@ class CleanCasesMainView(BaseProjectDataView):
     use_bootstrap5,
     toggles.DATA_CLEANING_CASES.required_decorator(),
 ], name='dispatch')
-class CleanCasesSessionView(BaseProjectDataView):
+class CleanCasesSessionView(BulkEditSessionViewMixin, BaseProjectDataView):
     page_title = gettext_lazy("Clean Case Type")
     urlname = "data_cleaning_cases_session"
     template_name = "data_cleaning/clean_cases_session.html"
@@ -52,12 +53,9 @@ class CleanCasesSessionView(BaseProjectDataView):
             return redirect(reverse(CleanCasesMainView.urlname, args=(self.domain, )))
 
     @property
-    def session_id(self):
-        return self.kwargs['session_id']
-
-    @property
     @memoized
     def session(self):
+        # overriding mixin so that DoesNotExist can be raised in self.get() and we can redirect
         return BulkEditSession.objects.get(session_id=self.session_id)
 
     @property
